third_party/blink/web_tests/platform/generic/fast/dom/Range/range-on-detached-node-expected.txt

PASS range.selectNode(sample.firstChild); range.startContainer is sample
PASS range.setEndBefore(sample.firstChild); range.startContainer is sample
PASS range.setEndAfter(sample.firstChild); range.startContainer is sample
PASS range.setStartAfter(sample.firstChild); range.startContainer is sample
PASS range.setStartBefore(sample.firstChild); range.startContainer is sample
PASS range.selectNode(sample); range.startContainer threw exception InvalidNodeTypeError: Failed to execute 'selectNode' on 'Range': the given Node has no parent..
PASS range.setEndAfter(sample); range.startContainer threw exception InvalidNodeTypeError: Failed to execute 'setEndAfter' on 'Range': the given Node has no parent..
PASS range.setEndBefore(sample); range.startContainer threw exception InvalidNodeTypeError: Failed to execute 'setEndBefore' on 'Range': the given Node has no parent..
PASS range.setStartAfter(sample); range.startContainer threw exception InvalidNodeTypeError: Failed to execute 'setStartAfter' on 'Range': the given Node has no parent..
PASS range.setStartBefore(sample); range.startContainer threw exception InvalidNodeTypeError: Failed to execute 'setStartBefore' on 'Range': the given Node has no parent..
PASS successfullyParsed is true

TEST COMPLETE
